requirements.txt
# ========================================
# DEPENDENCIAS DEL PROYECTO
# Sistema de Ofertas Laborales - Tacna
# ========================================

# Framework Web
Flask==2.3.3
Werkzeug==2.3.7

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Base de Datos NoSQL (MongoDB - NUEVO)
pymongo==4.6.1
dnspython==2.4.2  # Para conexiones MongoDB Atlas

# Utilidades
python-dotenv==1.0.0

# Opcional: Para mejorar el scraping (requieren compiladores C/C++)
# selenium==4.16.0  # Para sitios con JavaScript dinámico
# cloudscraper==1.2.71  # Para bypasear Cloudflare si es necesario
# lxml==4.9.3  # Parser XML/HTML más rápido (requiere compiladores)
